% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Append elements to sticky object}
\usage{
append(x, ...)
}
\arguments{
\item{x}{sticky object}

\item{...}{additional arguments to \code{base::append}

\code{append} preserves object attributes by wrapping \code{base::append} and
copying objects to the result.}
}
\description{
Append elements to sticky object
}
\seealso{
\code{\link[base]{append}}
}

