% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract1.R
\name{[.sticky}
\alias{[.sticky}
\title{Extract}
\usage{
\method{[}{sticky}(x, ...)
}
\arguments{
\item{x}{sticky; object to extract from}

\item{...}{additional parameters}
}
\description{
Extract
}
\details{
-tk
}

