% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.init}
\alias{.init}
\alias{.deinit}
\title{Internal, non-exported unit functions}
\usage{
.init(x)

.deinit(x)
}
\arguments{
\item{x}{vector}
}
\description{
These are internal functions for the sticky package. They are not
exported and should not be used directly. Instead use the proper interface,
\code{\link{sticky}} and \code{\link{unstick}}
}
\details{
\code{.init} : initializes the class
\code{.deinit} : uninitialize the class

These are internals functions not to be called directly, please use the
interfaces, \code{sticky} and \code{unsticky}.
}
\examples{
  # - tk

}
\seealso{
\code{sticky} and \code{unsticky}
}
