% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{.init}
\alias{.init}
\title{Internal, non-exported unit functions}
\usage{
.init(x)
}
\arguments{
\item{x}{vector}
}
\description{
These are internal functions for the sticky.attributes package. They are not
exported and should not be used directly. Instead use the proper interface,
\code{\link{sticky}}
}
\details{
\code{.init} : initializes the class
}

