% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.sticky}
\alias{subset}
\alias{subset.sticky}
\title{Subset elements to sticky object}
\usage{
\method{subset}{sticky}(x, ...)
}
\arguments{
\item{x}{sticky object}

\item{...}{additional arguments to \code{base::subset}

\code{subset} preserves object attributes by wrapping \code{base::subset} and
copying objects to the result.}
}
\description{
Subset elements to sticky object
}
\seealso{
\code{\link[base]{subset}}
}

